# swagger.yml
openapi: 3.0.0
info:
  title: PROCURE API Documentation
  version: 1.0.0
  description: Full API documentation for the PROCURE platform.
servers:
  - url: http://localhost:5000
  - url: https://your-domain.com
...
paths:
...
  /api/admin/suppliers:
    post:
      tags: [Suppliers]
      summary: Create Basic Supplier
      requestBody:
        content:
          application/json:
            schema:
              example:
                firstName: Tata
                lastName: Group
                email: tata@group.com
      responses:
        '201': { description: Created }

    get:
      tags: [Suppliers]
      summary: Get All Suppliers
      parameters:
        - in: query
          name: page
        - in: query
          name: limit
      responses:
        '200': { description: Supplier list }
  /api/admin/suppliers/suppliers:
    post:
      tags: [Suppliers]
      summary: Create Supplier with Logo
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                logo:
                  type: string
                  format: binary
                firstName:
                  type: string
      responses:
        '201': { description: Created }
  /api/admin/suppliers/{id}:
    get:
      tags: [Suppliers]
      summary: Get Supplier by ID
      parameters:
        - in: path
          name: id
      responses:
        '200': { description: Supplier }

    put:
      tags: [Suppliers]
      summary: Update Supplier
      requestBody:
        content:
          application/json:
            schema:
              example:
                firstName: Updated
      responses:
        '200': { description: Updated }

    delete:
      tags: [Suppliers]
      summary: Delete Supplier
      responses:
        '200': { description: Deleted }
...